@JackStouffer 2016-07-11
One benefit I forgot to mention is that this will also speed up `std.conv.to` for strings considerably.