@JackStouffer 2016-07-18
Ping.

This is a significant speed increase for a very common function.