@9il 2016-07-22
Thanks! Please add or refer in PR description a unittest that cover UTF (also note about UTF). Also please replace `Unqual!(typeof(s.front)) c;` with `uint c`, this will speed up code for some rare cases