@9il 2016-07-22
We need more such optimisations