@JackStouffer 2016-07-24
> Please add or refer in PR description a unittest that cover UTF (also note about UTF).

What exactly are you looking for? Parse already ignores everything that's not a number.