@wilzbach 2016-08-31
@JackStouffer FYI attached [issue 16451](https://issues.dlang.org/show_bug.cgi?id=16451) to this PR ;-)